+++ src/components/UserMenuBtn/UserMenuBtn.vue
@@ -6,6 +6,7 @@
 	import { mdiAccount, mdiLoginVariant } from '@mdi/js'
 	import useCustomizableOptions, { type CustomizableOptions } from '@/composables/useCustomizableOptions'
 	import { defaultOptions } from './config'
+	import SyIcon from '@/components/Customs/SyIcon/SyIcon.vue'
 
 	type MenuItem = { text: string, value: string, link?: string, to?: RouteLocationRaw }
 
@@ -55,14 +56,14 @@
 		class="user-menu-btn"
 	>
 		<template #append-icon>
-			<VIcon
+			<SyIcon
 				v-if="!hideUserIcon"
+				decorative
+				:icon="mdiAccount"
 				:size="isMobileView ? 'x-large' : 'default'"
 				class="vd-user-icon mr-0 pa-2"
 				v-bind="options['icon']"
-			>
-				{{ mdiAccount }}
-			</VIcon>
+			/>
 		</template>
 		<template #footer-list-item>
 			<slot>
@@ -73,10 +74,11 @@
 					@click="$emit('logout')"
 				>
 					<div class="d-flex">
-						<VIcon
+						<SyIcon
 							:icon="mdiLoginVariant"
 							class="mr-4"
 							v-bind="options['logoutIcon']"
+							decorative
 						/>
 						<VListItemTitle class="logout">
 							{{ props.logoutText }}